fix: compaction filter non-static Change
Presently the compaction filter declare the Change case of the Decision enum as
holding 'static [u8] slice. That made it rather useless in safe Rust as any
non-trivial data change must leak the returned data. Due to the way RocksDB
exposes the filter functionality through C API for the ordinary compaction
filter callback using 'static for data is inevitable.

However for a filter instance created via a compaction filter factory 'static is
not necessary. RocksDB guarantees that such filter instance will be accessed
from the single thread and the change data pointer returned by filter will not
outlive the instance. Thus is is safe to return Change holding a slice reference
bound by the lifetime of the filter instance.

This change fixes this via introducing lifetime bound parameter for the Decision
enum. Then it requires that the CompactionFilter::filter() trait method return
Decision bound by the filter instance, not 'static.

This is a breaking change due to the new lifetime parameter on a public struct.
However upgrading the existing code should be straitforward via adding <'static>
as necessary to the CompactionDecision type.